+++ NEWS.md
@@ -5,6 +5,7 @@
 * `ex_disagg` is now an exported data object. It contains an example `knnst` object for use in examples and tests.
 * Changed how unnamed input `mon_flow` is treated. Now the columns (sites) are renamed in `knn_space_time_disagg()`, instead of in `as.data.frame.knnst()`. #33
 * Added computation of (`knnst_spatial_cor()`) and plotting for (`plot.knnst_spcor()`) spatial correlation between sites. #29
+* Added computation of (`knnst_temporal_cor()`) and plotting for (`plot.knnst_tmpcor()`) temporal correlation between months at a site. #30
 
 # knnstdisagg 0.1.0